+++ cobs/query/search.hpp
@@ -39,7 +39,7 @@ public:
     virtual void search(
         const std::string& query,
         std::vector<SearchResult>& result,
-        double threshold = 0.0, size_t num_results = 0) = 0;
+        double threshold = 0.0, uint64_t num_results = 0) = 0;
 
 public:
     //! timer of different query phases
